Occupier - Poorly written, not funny yet.
Not funny yet, still just insulting.  I intend this guy to be a doofus
follower of fashion, not flame bait.

NOTE: Other stories will be about a bond trader, an oil lobbyist, and
several other stereotypes.  Occupy Guy is just the first, and I don't
mean to pick on them in particular.